Module type Bap_api.S

Language processor interface.

type t

val language : string

language a name of a language, e.g., "C"

val parse : ( api -> filename option ) -> api list -> t Core_kernel.Or_error.t

parse get_api apis creates a language processor from a list of api. Function get_api api must return a name of an existing file, that corresponds to the given api. The apis parameter is a list of api names.

val mapper : t -> Bap.Std.Term.mapper

the processor itself